Fix models filse structure
