+++ src-tauri/src/commands/mod.rs
@@ -1,3 +1,3 @@
-pub(crate) mod data;
+pub(crate) mod books;
 pub(crate) mod scanner;
 pub(crate) mod settings;
