+++ src-tauri/src/models/book.rs
@@ -7,13 +7,6 @@ use diesel::Queryable;
 use serde::{Deserialize, Serialize};
 
 #[derive(Queryable, Insertable, Serialize, Deserialize, Debug)]
-#[table_name = "authors"]
-pub struct Author {
-    pub name: String,
-    pub relative_img_path: Option<String>,
-}
-
-#[derive(Queryable, Insertable, Serialize, Deserialize, Debug)]
 #[table_name = "books"]
 pub struct Book {
     pub title: String,
@@ -24,7 +17,6 @@ pub struct Book {
     pub create_date: Option<NaiveDateTime>,
 }
 
-// BooksRead table model
 #[derive(Queryable, Insertable, Serialize, Deserialize, Debug)]
 #[table_name = "books_read"]
 pub struct BookRead {
@@ -34,20 +26,4 @@ pub struct BookRead {
     pub tier: Option<i32>,
     pub note: Option<String>,
     pub read_date: Option<NaiveDateTime>,
-}
-
-// Tags table model
-#[derive(Queryable, Insertable, Serialize, Deserialize, Debug)]
-#[table_name = "tags"]
-pub struct Tag {
-    pub id: i32,
-    pub name: String,
-}
-
-// TagsBooks table model
-#[derive(Queryable, Insertable, Serialize, Deserialize, Debug)]
-#[table_name = "tags_books"]
-pub struct TagBook {
-    pub tag_id: i32,
-    pub book_title: String,
 }
