+++ src-tauri/src/models/mod.rs
@@ -1,2 +1,4 @@
-pub(crate) mod models;
+pub(crate) mod author;
+pub(crate) mod book;
 pub(crate) mod query;
+pub(crate) mod tag;
